double type changed to Double reviewed by @Homework

+++ backend/src/main/java/hu/flowacademy/meteo/model/HourlyData.java
@@ -23,28 +23,28 @@ public class HourlyData {
     @Id
     @GeneratedValue
     @JsonIgnore
-    private int id;
+    private long id;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy.MM.dd HH:mm", timezone = "UTC")
     private Date date;
-    private double airHumidity;
-    private double airPressure;
-    private double windSpeed;
-    private double solarCellChargingVoltage;
-    private double externalBatteryVoltage;
-    private double irradiation;
-    private double freeze;
-    private double rain;
-    private double windDirection;
-    private double windGust;
-    private double soilMoisture90cm;
-    private double leafMoisture;
-    private double soilTemperature0cm;
-    private double airTemperature;
-    private double internalBatteryVoltage;
-    private double soilMoisture30cm;
-    private double soilMoisture60cm;
-    private double lightUnit;
-    private double soilMoisture120cm;
-    private double precipitationCounter;
+    private Double airHumidity;
+    private Double airPressure;
+    private Double windSpeed;
+    private Double solarCellChargingVoltage;
+    private Double externalBatteryVoltage;
+    private Double irradiation;
+    private Double freeze;
+    private Double rain;
+    private Double windDirection;
+    private Double windGust;
+    private Double soilMoisture90cm;
+    private Double leafMoisture;
+    private Double soilTemperature0cm;
+    private Double airTemperature;
+    private Double internalBatteryVoltage;
+    private Double soilMoisture30cm;
+    private Double soilMoisture60cm;
+    private Double lightUnit;
+    private Double soilMoisture120cm;
+    private Double precipitationCounter;
 }
